añade ejercicios y presentacion pdf

+++ Rmd/build_slides.R
@@ -8,9 +8,13 @@ output_dir <- "./slides"
 # 2. Convert them to PDF using `chrome_print()`
 # 3. Move them to `output_dir`
 for (file in list.files(path = input_dir, pattern = ".Rmd", recursive = TRUE)) {
-  input <- file.path(input_dir, file)
+  input  <- file.path(input_dir, file)
   output <- paste0(file_path_sans_ext(input), ".pdf")
-  pagedown::chrome_print(input)
-  file.copy(output, output_dir)
-  file.remove(output)
+  target <- file.path(output_dir, paste0(file_path_sans_ext(basename(file)), ".pdf"))
+  
+  if (!file.exists(target)) {
+    pagedown::chrome_print(input)
+    file.copy(output, output_dir)
+    file.remove(output)
+  }
 }
